Take webpackOptions.module.rules into account for style loaders, when using webpack 2.x
As of webpack 2.x, loaders are now specified as rules:
When an inline style is used with less (style.less {...}) The marko-loader used to only search through options.module.loaders for the less-loader.
In order to fix this, and make the fix work for previous versions of webpack (viz, 1.x), webpack/lib/ResultSet.js has been copied into this module.
Fixes #4